requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements_dev.txt requirements/dev.in requirements/production.in
#
ansible==5.1.0
    # via -r requirements/dev.in
ansible-core==2.12.1
    # via ansible
arrow==1.2.1
    # via django-q
asgiref==3.4.1
    # via django
black==21.12b0
    # via -r requirements/dev.in
blessed==1.19.0
    # via django-q
brotli==1.0.9
    # via django-compress-staticfiles
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via black
coverage==6.2
    # via -r requirements/dev.in
cryptography==36.0.1
    # via
    #   ansible-core
    #   pyjwt
csscompressor==0.9.5
    # via django-compress-staticfiles
defusedxml==0.7.1
    # via python3-openid
distlib==0.3.4
    # via virtualenv
dj-rest-auth==2.2.1
    # via -r requirements/production.in
django==3.2.10
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
    #   django-allauth
    #   django-compress-staticfiles
    #   django-cors-headers
    #   django-picklefield
    #   django-q
    #   djangorestframework
django-allauth==0.47.0
    # via -r requirements/production.in
django-compress-staticfiles==1.0.1b0
    # via -r requirements/production.in
django-cors-headers==3.10.1
    # via -r requirements/production.in
django-picklefield==3.0.1
    # via django-q
django-q==1.3.9
    # via -r requirements/production.in
djangorestframework==3.13.1
    # via
    #   -r requirements/production.in
    #   dj-rest-auth
filelock==3.4.2
    # via virtualenv
gunicorn==20.1.0
    # via -r requirements/production.in
identify==2.4.1
    # via pre-commit
idna==3.3
    # via requests
jinja2==3.0.3
    # via ansible-core
markupsafe==2.0.1
    # via jinja2
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
oauthlib==3.1.1
    # via requests-oauthlib
packaging==21.3
    # via ansible-core
pathspec==0.9.0
    # via black
platformdirs==2.4.1
    # via
    #   black
    #   virtualenv
pre-commit==2.16.0
    # via -r requirements/dev.in
psycopg2-binary==2.9.3
    # via -r requirements/production.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
pyparsing==3.0.6
    # via packaging
python-dateutil==2.8.2
    # via
    #   -r requirements/production.in
    #   arrow
python-dotenv==0.19.2
    # via -r requirements/production.in
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   ansible-core
    #   pre-commit
redis==3.5.3
    # via django-q
requests==2.26.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
resolvelib==0.5.4
    # via ansible-core
rjsmin==1.2.0
    # via django-compress-staticfiles
six==1.16.0
    # via
    #   blessed
    #   python-dateutil
    #   virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==1.2.3
    # via black
typing-extensions==4.0.1
    # via black
urllib3==1.26.7
    # via
    #   -r requirements/production.in
    #   requests
virtualenv==20.12.1
    # via pre-commit
wcwidth==0.2.5
    # via blessed
zxcvbn==4.4.28
    # via -r requirements/production.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
